requirements.txt
# Core
aiogram==3.17.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# HTTP requests
requests==2.32.3

# PDF processing
PyMuPDF==1.25.2

# OCR
easyocr==1.7.1
pillow==11.0.0

# Word documents
python-docx==1.1.2

# Async file operations
aiofiles==24.1.0